updated C code and vertices for correct untimed execution.
