+++ c_code/comms_s.c
@@ -1,11 +1,15 @@
 // SpiNNaker API
 #include "spin1_api.h"
 
+// graph-front-end
+#include <simulation.h>
+
 // mlp
 #include "mlp_params.h"
 #include "mlp_types.h"
 #include "mlp_externs.h"
 
+#include "init_s.h"
 #include "comms_s.h"
 #include "process_s.h"
 
@@ -42,7 +46,7 @@ void s_receivePacket (uint key, uint payload)
       sf_thrds_pend = 1;
 
       // and advance tick
-      spin1_schedule_callback (sf_advance_tick, NULL, NULL, SPINN_S_TICK_P);
+      spin1_schedule_callback (sf_advance_tick, 0, 0, SPINN_S_TICK_P);
     }
     else
     {
@@ -62,7 +66,7 @@ void s_receivePacket (uint key, uint payload)
     #endif
 
       // stop timer ticks,
-      simulation_exit ();
+//lap      simulation_exit ();
 
       // report no error,
       done(SPINN_NO_ERROR);
@@ -77,7 +81,7 @@ void s_receivePacket (uint key, uint payload)
   if (new_tail == s_pkt_queue.head)
   {
       // stop timer ticks,
-      simulation_exit ();
+//lap      simulation_exit ();
 
       // report queue full error,
       done(SPINN_QUEUE_FULL);
@@ -96,7 +100,7 @@ void s_receivePacket (uint key, uint payload)
     if (!s_active)
     {
       s_active = TRUE;
-      spin1_schedule_callback (s_process, NULL, NULL, SPINN_S_PROCESS_P);
+      spin1_schedule_callback (s_process, 0, 0, SPINN_S_PROCESS_P);
     }
   }
 }
