+++ c_code/comms_w.c
@@ -1,11 +1,15 @@
 // SpiNNaker API
 #include "spin1_api.h"
 
+// graph-front-end
+#include <simulation.h>
+
 // mlp
 #include "mlp_params.h"
 #include "mlp_types.h"
 #include "mlp_externs.h"
 
+#include "init_w.h"
 #include "comms_w.h"
 #include "process_w.h"
 
@@ -111,7 +115,7 @@ void w_stopPacket (uint key)
       io_printf (IO_BUF, "wrp scheduling wf_advance_tick\n");
     #endif
 
-    spin1_schedule_callback (wf_advance_tick, NULL, NULL, SPINN_WF_TICK_P);
+    spin1_schedule_callback (wf_advance_tick, 0, 0, SPINN_WF_TICK_P);
   }
   else
   {
@@ -132,7 +136,7 @@ void w_networkStopPacket (void)
   #endif
 
     // stop timer ticks,
-    simulation_exit ();
+//lap    simulation_exit ();
 
     // report no error,
     done(SPINN_NO_ERROR);
@@ -167,7 +171,7 @@ void w_ldsrPacket (uint payload)
 
     // and advance tick
     //TODO: check if need to schedule or can simply call
-    spin1_schedule_callback (wb_advance_tick, NULL, NULL, SPINN_WB_TICK_P);
+    spin1_schedule_callback (wb_advance_tick, 0, 0, SPINN_WB_TICK_P);
   }
   else
   {
@@ -224,7 +228,7 @@ void w_forwardPacket (uint key, uint payload)
         io_printf (IO_BUF, "wfpkt scheduling wf_advance_tick\n");
       #endif
 
-      spin1_schedule_callback (wf_advance_tick, NULL, NULL, SPINN_WF_TICK_P);
+      spin1_schedule_callback (wf_advance_tick, 0, 0, SPINN_WF_TICK_P);
     }
     else
     {
@@ -253,7 +257,7 @@ void w_backpropPacket (uint key, uint payload)
   if (new_tail == w_delta_pkt_q.head)
   {
       // stop timer ticks,
-      simulation_exit ();
+//lap      simulation_exit ();
 
       // report queue full error,
       done(SPINN_QUEUE_FULL);
@@ -273,7 +277,7 @@ void w_backpropPacket (uint key, uint payload)
     if (!wb_active)
     {
       wb_active = TRUE;
-      spin1_schedule_callback (wb_process, NULL, NULL, SPINN_WB_PROCESS_P);
+      spin1_schedule_callback (wb_process, 0, 0, SPINN_WB_PROCESS_P);
     }
   }
 }
