+++ c_code/init_i.h
@@ -2,5 +2,6 @@
 #define __INIT_I_H__
 
 uint i_init (void);
+void done (uint ec);
 
 #endif
