+++ c_code/init_t.h
@@ -2,6 +2,7 @@
 #define __INIT_T_H__
 
 uint t_init (void);
+void done (uint ec);
 
 #endif
