+++ c_code/init_w.h
@@ -2,6 +2,7 @@
 #define __INIT_W_H__
 
 uint w_init (void);
+void done (uint ec);
 
 #endif
