+++ c_code/process_t.c
@@ -1,6 +1,9 @@
 // SpiNNaker API
 #include "spin1_api.h"
 
+// graph-front-end
+#include <simulation.h>
+
 // mlp
 #include "mlp_params.h"
 #include "mlp_types.h"
@@ -22,6 +25,9 @@ void tf_process (uint null0, uint null1)
 #ifdef TRACE
   io_printf (IO_BUF, "tb_process\n");
 #endif
+
+  (void) null0;
+  (void) null1;
 
   // process packet queue
   // access queue with interrupts disabled
@@ -119,14 +125,14 @@ void tf_process (uint null0, uint null1)
 
 	  // send stop criterion packet,
 	  //TODO: check if need to schedule or can simply call
-	  tf_send_stop (NULL, NULL);
+	  tf_send_stop (0, 0);
 
 	  // and advance tick if last group
 	  //NOTE: last group in the chain does not get a stop decision
 	  if (tcfg.is_last_output_group)
 	  {
             //TODO: check if need to schedule or can simply call
-            tf_advance_tick (NULL, NULL);
+            tf_advance_tick (0, 0);
           }
 	}
 	else
@@ -151,7 +157,7 @@ void tf_process (uint null0, uint null1)
 
           // and advance tick
           //TODO: check if need to schedule or can simply call
-          tf_advance_tick (NULL, NULL);
+          tf_advance_tick (0, 0);
         }
         else
         {
@@ -185,6 +191,9 @@ void tb_process (uint null0, uint null1)
 #ifdef TRACE
   io_printf (IO_BUF, "tb_process\n");
 #endif
+
+  (void) null0;
+  (void) null1;
 
   // compute deltas based on pre-computed errors,
   //TODO: this needs checking!
@@ -255,7 +264,7 @@ void tb_process (uint null0, uint null1)
     io_printf (IO_BUF, "tbp calling tb_advance_tick\n");
 #endif
 
-    tb_advance_tick (NULL, NULL);
+    tb_advance_tick (0, 0);
   }
   else
   {
@@ -290,6 +299,9 @@ void tf_advance_tick (uint null0, uint null1)
 #if SPINN_OUTPUT_HISTORY == TRUE
   //TODO: dump outputs to SDRAM for record keeping,
 #endif
+
+  (void) null0;
+  (void) null1;
 
   // check if done with event
   if (tick_stop)
@@ -332,6 +344,9 @@ void tb_advance_tick (uint null0, uint null1)
   io_printf (IO_BUF, "tb_tick: %d/%d\n", tick, tot_tick);
 #endif
 
+  (void) null0;
+  (void) null1;
+
   // update pointer to processing unit outputs,
   tb_procs = 1 - tb_procs;
 
@@ -362,7 +377,7 @@ void tb_advance_tick (uint null0, uint null1)
     tick--;
 
     // and trigger computation
-    spin1_schedule_callback (tb_process, NULL, NULL, SPINN_TB_PROCESS_P);
+    spin1_schedule_callback (tb_process, 0, 0, SPINN_TB_PROCESS_P);
   }
 }
 // ------------------------------------------------------------------------
@@ -479,7 +494,7 @@ void t_advance_example (void)
 #endif
 
         // stop timer ticks,
-        simulation_exit ();
+//lap        simulation_exit ();
 
         // report no error,
         done(SPINN_NO_ERROR);
@@ -500,7 +515,7 @@ void t_advance_example (void)
       }
 
       // stop timer ticks,
-      simulation_exit ();
+//lap      simulation_exit ();
 
       // report no error,
       done(SPINN_NO_ERROR);
@@ -575,7 +590,7 @@ void t_advance_example (void)
 
     // and send unit outputs to w cores
     //TODO: check if need to schedule or can simply call
-    t_init_outputs (NULL, NULL);
+    t_init_outputs (0, 0);
   }
   else
   {
@@ -610,7 +625,7 @@ void t_switch_to_fw (void)
   if (t_net_pkt_q.head != t_net_pkt_q.tail)
   {
     // if queue not empty schedule FORWARD processing
-    spin1_schedule_callback (tf_process, NULL, NULL, SPINN_TF_PROCESS_P);
+    spin1_schedule_callback (tf_process, 0, 0, SPINN_TF_PROCESS_P);
   }
   else
   {
@@ -648,7 +663,7 @@ void t_switch_to_bp (void)
 
   // start processing in BACKPROP phase,
   //TODO: check!
-  spin1_schedule_callback (tb_process, NULL, NULL, SPINN_TB_PROCESS_P);
+  spin1_schedule_callback (tb_process, 0, 0, SPINN_TB_PROCESS_P);
 
   // and restore interrupts
   spin1_mode_restore (cpsr);
@@ -667,6 +682,9 @@ void tf_send_stop (uint null0, uint null1)
 #ifdef TRACE
   io_printf (IO_BUF, "tf_send_stop\n");
 #endif
+
+  (void) null0;
+  (void) null1;
 
   // "aggregate" criteria,
   tf_stop_crit = tf_stop_crit && tf_chain_prev;
@@ -722,7 +740,10 @@ void t_init_outputs (uint null0, uint null1)
   io_printf (IO_BUF, "t_init_outputs\n");
 #endif
 
-  // initialize every unit output and send for processing
+  (void) null0;
+  (void) null1;
+
+  // initialise every unit output and send for processing
   for (uint i = 0; i < tcfg.num_units; i++)
   {
     // setup the initial output value.
@@ -746,7 +767,7 @@ void t_init_outputs (uint null0, uint null1)
     }
 
 #ifdef DEBUG_CFG3
-    io_printf (IO_BUF, "to[%u]: 0x%08x\n", i, t_outputs[i]);
+  io_printf (IO_BUF, "to[%u]: 0x%08x\n", i, t_outputs[i]);
 #endif
 
     // and send unit output to weight cores
@@ -757,8 +778,8 @@ void t_init_outputs (uint null0, uint null1)
           );
 
 #ifdef DEBUG
-    pkt_sent++;
-    sent_fwd++;
+  pkt_sent++;
+  sent_fwd++;
 #endif
   }
 }
@@ -1045,9 +1066,9 @@ void out_bias (uint inx)
 // ------------------------------------------------------------------------
 void out_weak_clamp (uint inx)
 {
-  #ifdef TRACE_VRB
-    io_printf (IO_BUF, "out_weak_clamp\n");
-  #endif
+#ifdef TRACE_VRB
+  io_printf (IO_BUF, "out_weak_clamp\n");
+#endif
 
   // TODO: if training, store the injected value in SDRAM. This memory area needs
   // to be allocated during initialization
@@ -1190,6 +1211,8 @@ void out_hard_clamp_back (uint inx)
   io_printf (IO_BUF, "out_hard_clamp_back\n");
 #endif
 
+  (void) inx;
+
 /*
   short_activ_t * tmp = t_out_hard_clamp_data + tick * tcfg.num_units;
 
@@ -1208,6 +1231,8 @@ void out_weak_clamp_back (uint inx)
 #ifdef TRACE_VRB
   io_printf (IO_BUF, "out_weak_clamp_back\n");
 #endif
+
+  (void) inx;
 }
 // ------------------------------------------------------------------------
