+++ c_code/process_w.c
@@ -1,6 +1,9 @@
 // SpiNNaker API
 #include "spin1_api.h"
 
+// graph-front-end
+#include <simulation.h>
+
 // mlp
 #include "mlp_params.h"
 #include "mlp_types.h"
@@ -22,6 +25,9 @@ void wf_process (uint null0, uint null1)
 #ifdef TRACE
   io_printf (IO_BUF, "wf_process\n");
 #endif
+
+  (void) null0;
+  (void) null1;
 
   // compute all net block dot-products and send them for accumulation,
   for (uint j = 0; j < wcfg.num_cols; j++)
@@ -80,7 +86,7 @@ void wf_process (uint null0, uint null1)
     io_printf (IO_BUF, "wfp calling wf_advance_tick\n");
 #endif
 
-    wf_advance_tick (NULL, NULL);
+    wf_advance_tick (0, 0);
   }
   else
   {
@@ -106,6 +112,9 @@ void wb_process (uint null0, uint null1)
 #ifdef PROFILE
   io_printf (IO_STD, "tin:  %u\n", tc[T2_COUNT]);
 #endif
+
+  (void) null0;
+  (void) null1;
 
   // process delta packet queue
   // access queue with interrupts disabled
@@ -259,7 +268,7 @@ void wb_process (uint null0, uint null1)
 #endif
 
         //TODO: check if need to schedule or can simply call
-        wb_advance_tick (NULL, NULL);
+        wb_advance_tick (0, 0);
       }
       else
       {
@@ -718,6 +727,9 @@ void wf_advance_tick (uint null0, uint null1)
   }
 #endif
 
+  (void) null0;
+  (void) null1;
+
   // change packet key colour,
   fwdKey ^= SPINN_COLOUR_KEY;
 
@@ -735,7 +747,7 @@ void wf_advance_tick (uint null0, uint null1)
     tick++;
 
     // and trigger computation
-    spin1_schedule_callback (wf_process, NULL, NULL, SPINN_WF_PROCESS_P);
+    spin1_schedule_callback (wf_process, 0, 0, SPINN_WF_PROCESS_P);
   }
 }
 // ------------------------------------------------------------------------
@@ -762,6 +774,9 @@ void wb_advance_tick (uint null0, uint null1)
 #ifdef DEBUG_VRB
   io_printf (IO_BUF, "wb: num_ticks: %d, tick: %d\n", num_ticks, tick);
 #endif
+
+  (void) null0;
+  (void) null1;
 
   // change packet key colour,
   bkpKey ^= SPINN_COLOUR_KEY;
@@ -840,7 +855,7 @@ void wf_advance_event (void)
     tick++;
 
     // and trigger computation
-    spin1_schedule_callback (wf_process, NULL, NULL, SPINN_WF_PROCESS_P);
+    spin1_schedule_callback (wf_process, 0, 0, SPINN_WF_PROCESS_P);
   }
 }
 // ------------------------------------------------------------------------
@@ -875,7 +890,7 @@ void w_advance_example (void)
     if (++epoch >= ncfg.num_epochs)
     {
         // stop timer ticks,
-        simulation_exit ();
+//lap        simulation_exit ();
 
         // report no error,
         done(SPINN_NO_ERROR);
