+++ c_code/threshold.c
@@ -2,6 +2,7 @@
 #include "spin1_api.h"
 
 // graph-front-end
+#include "common-typedefs.h"
 #include <data_specification.h>
 #include <simulation.h>
 
@@ -208,29 +209,31 @@ uint init ()
   io_printf (IO_BUF, "threshold\n");
 
   // read the data specification header
-  address_t data_address = data_specification_get_data_address ();
-  if (!data_specification_read_header (data_address)) {
+  data_specification_metadata_t * data =
+          data_specification_get_data_address();
+  if (!data_specification_read_header (data))
+  {
 	  return (SPINN_CFG_UNAVAIL);
   }
 
   // set up the simulation interface (system region)
   //NOTE: these variables are not used!
-  uint simulation_ticks, infinite_run, time, timer_period;
-  if (!simulation_initialise(
-          data_specification_get_region(SYSTEM, data_address),
-          APPLICATION_NAME_HASH, &timer_period, &simulation_ticks,
-          &infinite_run, &time, 0, 0)) {
-      return (SPINN_CFG_UNAVAIL);
+  uint32_t n_steps, run_forever, step;
+  if (!simulation_steps_initialise(
+      data_specification_get_region(SYSTEM, data),
+      APPLICATION_NAME_HASH, &n_steps, &run_forever, &step, 0, 0))
+  {
+    return (SPINN_CFG_UNAVAIL);
   }
 
   // network configuration address
-  address_t nt = data_specification_get_region (NETWORK, data_address);
+  address_t nt = data_specification_get_region (NETWORK, data);
 
   // initialise network configuration from SDRAM
   spin1_memcpy (&ncfg, nt, sizeof(network_conf_t));
 
   // core configuration address
-  address_t dt = data_specification_get_region (CORE, data_address);
+  address_t dt = data_specification_get_region (CORE, data);
 
   // initialise core-specific configuration from SDRAM
   spin1_memcpy (&tcfg, dt, sizeof(t_conf_t));
@@ -239,19 +242,19 @@ uint init ()
   if (tcfg.input_grp)
   {
     it = (activation_t *) data_specification_get_region
-		  (INPUTS, data_address);
+		  (INPUTS, data);
   }
 
   // targets
   if (tcfg.output_grp)
   {
     tt = (activation_t *) data_specification_get_region
-		  (TARGETS, data_address);
+		  (TARGETS, data);
   }
 
   // example set
   es = (struct mlp_set *) data_specification_get_region
-		  (EXAMPLE_SET, data_address);
+		  (EXAMPLE_SET, data);
 
 #ifdef DEBUG_CFG5
   io_printf (IO_BUF, "ne: %u\n", es->num_examples);
@@ -263,7 +266,7 @@ uint init ()
 
   // examples
   ex = (struct mlp_example *) data_specification_get_region
-		  (EXAMPLES, data_address);
+		  (EXAMPLES, data);
 
 #ifdef DEBUG_CFG5
   for (uint i = 0; i < es->num_examples; i++)
@@ -277,7 +280,7 @@ uint init ()
 
   // events
   ev = (struct mlp_event *) data_specification_get_region
-		  (EVENTS, data_address);
+		  (EVENTS, data);
 
 #ifdef DEBUG_CFG5
   uint evi = 0;
@@ -296,7 +299,7 @@ uint init ()
 
   // routing keys
   rt = (uint *) data_specification_get_region
-		  (ROUTING, data_address);
+		  (ROUTING, data);
 
 #ifdef DEBUG_CFG0
   io_printf (IO_BUF, "og: %d\n", tcfg.output_grp);
@@ -349,125 +352,25 @@ uint init ()
 
 
 // ------------------------------------------------------------------------
-// check exit code and print details of the state
-// ------------------------------------------------------------------------
-void done (uint ec)
-{
-  // report problems -- if any
-  switch (ec)
-  {
-    case SPINN_NO_ERROR:
-      io_printf (IO_BUF, "simulation OK\n");
-      break;
-
-    case SPINN_CFG_UNAVAIL:
-      io_printf (IO_BUF, "core configuration failed\n");
-      io_printf(IO_BUF, "simulation aborted\n");
-      break;
-
-    case SPINN_QUEUE_FULL:
-      io_printf (IO_BUF, "packet queue full\n");
-      io_printf(IO_BUF, "simulation aborted\n");
-      break;
-
-    case SPINN_MEM_UNAVAIL:
-      io_printf (IO_BUF, "malloc failed\n");
-      io_printf(IO_BUF, "simulation aborted\n");
-      break;
-
-    case SPINN_UNXPD_PKT:
-      io_printf (IO_BUF, "unexpected packet received - abort!\n");
-      io_printf(IO_BUF, "simulation aborted\n");
-      break;
-
-    case SPINN_TIMEOUT_EXIT:
-      io_printf (IO_BUF, "timeout (h:%u e:%u p:%u t:%u) - abort!\n",
-                 epoch, example, phase, tick
-                );
-      io_printf(IO_BUF, "simulation aborted\n");
-#ifdef DEBUG_TO
-      io_printf (IO_BUF, "(tactive:%u ta:%u/%u tb:%u/%u)\n",
-                  t_active, tf_arrived, tcfg.num_units,
-                  tb_arrived, tcfg.num_units
-                );
-      io_printf (IO_BUF, "(tsr:%u tsa:%u/%u)\n",
-                  t_sync_rdy, t_sync_arrived, tcfg.fwd_sync_expected
-                );
-      io_printf (IO_BUF, "(tcr:%u)\n",
-                  tf_chain_rdy
-                );
-#endif
-      if (tcfg.write_out)  // make sure the output monitor closes!
-      {
-        send_outputs_to_host (SPINN_HOST_FINAL, tick);
-      }
-      break;
-  }
-
-  // report diagnostics
-  #ifdef DEBUG
-    io_printf (IO_BUF, "total ticks:%d\n", tot_tick);
-    io_printf (IO_BUF, "total recv:%d\n", pkt_recv);
-    io_printf (IO_BUF, "total sent:%d\n", pkt_sent);
-    io_printf (IO_BUF, "recv: fwd:%d bkp:%d\n", recv_fwd, recv_bkp);
-    io_printf (IO_BUF, "sent: fwd:%d bkp:%d\n", sent_fwd, sent_bkp);
-    io_printf (IO_BUF, "sync recv:%d\n", spk_recv);
-    if (tcfg.is_first_output_group)
-    {
-      io_printf (IO_BUF, "chain recv: first\n");
-    }
-    else
-    {
-	  io_printf (IO_BUF, "chain recv:%d\n", chn_recv);
-    }
-    if (tcfg.is_last_output_group)
-    {
-      io_printf (IO_BUF, "stop sent:%d\n", stp_sent);
-      io_printf (IO_BUF, "stpn sent:%d\n", stn_sent);
-    }
-    else
-    {
-      io_printf (IO_BUF, "chain sent:%d\n", chn_sent);
-      io_printf (IO_BUF, "stop recv:%d\n", stp_recv);
-      io_printf (IO_BUF, "stpn recv:%d\n", stn_recv);
-    }
-    if (wrng_phs) io_printf (IO_BUF, "wrong phase:%d\n", wrng_phs);
-    if (wrng_tck) io_printf (IO_BUF, "wrong tick:%d\n", wrng_tck);
-    if (wrng_btk) io_printf (IO_BUF, "wrong btick:%d\n", wrng_btk);
-  #endif
-
-  io_printf (IO_BUF, "stopping simulation\n");
-  io_printf (IO_BUF, "-----------------------\n");
-
-  // and say goodbye
-  io_printf (IO_BUF, "<< mlp\n");
-}
-// ------------------------------------------------------------------------
-
-
-// ------------------------------------------------------------------------
 // timer callback: check that there has been progress in execution.
 // If no progress has been made terminate with SPINN_TIMEOUT_EXIT exit code.
 // ------------------------------------------------------------------------
 void timeout (uint ticks, uint null)
 {
-  if (ticks == 1)
-  {
-      spin1_schedule_callback (t_init_outputs, NULL, NULL, SPINN_T_INIT_OUT_P);
-      return;
-  }
+  (void) ticks;
+  (void) null;
 
   // check if progress has been made
   if ((to_epoch == epoch) && (to_example == example) && (to_tick == tick))
   {
-      // stop timer ticks,
-      simulation_exit();
+    // stop timer ticks,
+    simulation_exit();
 
-      // report timeout error,
-      done(SPINN_TIMEOUT_EXIT);
+    // report timeout error,
+    done(SPINN_TIMEOUT_EXIT);
 
-      // and let host know that we're ready
-      simulation_ready_to_read();
+    // and let host know that we're ready
+    simulation_ready_to_read();
   }
   else
   {
@@ -476,6 +379,24 @@ void timeout (uint ticks, uint null)
     to_example = example;
     to_tick    = tick;
   }
+}
+// ------------------------------------------------------------------------
+
+
+// ------------------------------------------------------------------------
+// start callback: get started by sending outputs to host and w cores.
+// ------------------------------------------------------------------------
+void get_started (void)
+{
+  // go,
+  io_printf (IO_BUF, "-----------------------\n");
+  io_printf (IO_BUF, "starting simulation\n");
+
+  if (tcfg.write_out)
+  {
+    spin1_schedule_callback (send_info_to_host, 0, 0, SPINN_T_SEND_OUTS_P);
+  }
+  spin1_schedule_callback (t_init_outputs, 0, 0, SPINN_T_INIT_OUT_P);
 }
 // ------------------------------------------------------------------------
 
@@ -504,47 +425,41 @@ void c_main ()
   }
 
   // set timer tick value (in microseconds),
-  spin1_set_timer_tick (SPINN_TIMER_TICK_PERIOD >> 1);
+//lap  spin1_set_timer_tick (SPINN_TIMER_TICK_PERIOD >> 1);
 
-  #ifdef PROFILE
-    // configure timer 2 for profiling
-    // enabled, 32 bit, free running, 16x pre-scaler
-    tc[T2_CONTROL] = SPINN_TIMER2_CONF;
-    tc[T2_LOAD] = SPINN_TIMER2_LOAD;
-  #endif
+#ifdef PROFILE
+  // configure timer 2 for profiling
+  // enabled, 32 bit, free running, 16x pre-scaler
+  tc[T2_CONTROL] = SPINN_TIMER2_CONF;
+  tc[T2_LOAD] = SPINN_TIMER2_LOAD;
+#endif
 
   // register callbacks,
   //NOTE: timeout escape -- in case something went wrong!
-  spin1_callback_on (TIMER_TICK, timeout, SPINN_TIMER_P);
+//lap  spin1_callback_on (TIMER_TICK, timeout, SPINN_TIMER_P);
 
   // packet received callbacks
   spin1_callback_on (MC_PACKET_RECEIVED, t_receivePacket, SPINN_PACKET_P);
   spin1_callback_on (MCPL_PACKET_RECEIVED, t_receivePacket, SPINN_PACKET_P);
 
-  // go,
-  io_printf (IO_BUF, "-----------------------\n");
-  io_printf (IO_BUF, "starting simulation\n");
+#ifdef PROFILE
+  uint start_time = tc[T2_COUNT];
+  io_printf (IO_BUF, "start count: %u\n", start_time);
+#endif
 
-  #ifdef PROFILE
-    uint start_time = tc[T2_COUNT];
-    io_printf (IO_BUF, "start count: %u\n", start_time);
-  #endif
+  // setup simulation,
+  simulation_set_start_function(get_started);
+  simulation_set_uses_timer(FALSE);
 
   // start execution,
   simulation_run();
 
-  #ifdef PROFILE
-    uint final_time = tc[T2_COUNT];
-    io_printf (IO_BUF, "final count: %u\n", final_time);
-    io_printf (IO_BUF, "execution time: %u us\n",
-                  (start_time - final_time) / SPINN_TIMER2_DIV);
-  #endif
-
-  // report results,
-  done (exit_code);
-
-  io_printf (IO_BUF, "stopping simulation\n");
-  io_printf (IO_BUF, "-----------------------\n");
+#ifdef PROFILE
+  uint final_time = tc[T2_COUNT];
+  io_printf (IO_BUF, "final count: %u\n", final_time);
+  io_printf (IO_BUF, "execution time: %u us\n",
+      (start_time - final_time) / SPINN_TIMER2_DIV);
+#endif
 
   // and say goodbye
   io_printf (IO_BUF, "<< mlp\n");
